ai-worker/requirements.txt
supabase
python-dotenv
requests
pillow